feat(readme): add beta release note for 0.6.2b1
We have pre-released the 0.6.2 beta version, which addresses numerous issues
reported in our logs. This commit updates the installation guide in the README to
include information on how to install this beta version. Users are advised that
this build has not undergone full QA testing and may contain issues. A
revert instruction to version 0.6.1 is also provided for users who encounter
problems.

BREAKING CHANGE: Installation instructions now include beta version
information. Users should be aware of potential issues with the0.6.2 beta
version and consider reverting to 0.6.1 if necessary.